requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
torchcodec>=0.2.0
numpy>=1.24.0
scipy>=1.10.0

# Configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Logging
comet-ml>=3.35.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Dataset handling
pandas>=2.0.0
gdown>=4.7.0
tqdm>=4.65.0

# Utilities
python-dotenv>=1.0.0
matplotlib>=3.7.0
tensorboard>=2.13.0

# Code quality
black>=23.0.0
isort>=5.12.0
